+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fixed `defaultPage` option to have no effect
 - Uniformed font and size for monospace views
 - Enlarged action buttons in `struct` and `structure` views
+- Changed rendering of after-annotations to render in a prelude when a value is expanded
 
 ## 1.0.0-beta.78 (21-09-2023)